requirements.txt
numpy==1.23.5
Pillow==10.2.0
streamlit==1.20.0
tensorflow==2.10.0
